switch to hpos-hc-connect crate WIP
