+++ src/lib.rs
@@ -2,23 +2,14 @@
 #![allow(clippy::unit_arg)]
 pub mod config;
 pub mod entries;
-pub mod websocket;
 use anyhow::Result;
-pub use websocket::{AdminWebsocket, AppWebsocket};
-pub mod host_zome_calls;
+pub use hpos_hc_connect::{AdminWebsocket, AppWebsocket};
 pub mod transaction_types;
-use host_zome_calls::get_all_published_hosted_happs;
-mod install_app;
-use install_app::install_holo_hosted_happs;
-mod uninstall_apps;
+mod utils;
+use utils::{get_pending_transactions, install_holo_hosted_happs, get_all_published_hosted_happs, uninstall_ineligible_happs, suspend_unpaid_happs};
 use tracing::{debug, error, info};
-use uninstall_apps::uninstall_ineligible_happs;
-mod suspend_happs;
-use suspend_happs::suspend_unpaid_happs;
 mod hbs;
 use hbs::{HbsClient, KycLevel};
-
-use crate::host_zome_calls::{get_pending_transactions, CoreAppClient};
 
 /// gets all the enabled happs from HHA
 /// installs and enables new happs that were registered by a provider and holochain disables those paused by provider in hha
