--- src/websocket.rs
@@ -1,142 +1,0 @@
-use anyhow::{anyhow, Context, Result};
-use holochain_conductor_api::{
-    AdminRequest, AdminResponse, AppInfo, AppRequest, AppResponse, AppStatusFilter, ZomeCall,
-};
-use holochain_types::app::InstalledAppId;
-use holochain_websocket::{connect, WebsocketConfig, WebsocketSender};
-use std::sync::Arc;
-use tracing::{info, trace};
-use url::Url;
-
-#[derive(Clone)]
-pub struct AdminWebsocket {
-    tx: WebsocketSender,
-}
-
-impl AdminWebsocket {
-    pub async fn connect(admin_port: u16) -> Result<Self> {
-        let url = format!("ws://localhost:{}/", admin_port);
-        let url = Url::parse(&url).context("invalid ws:// URL")?;
-        let websocket_config = Arc::new(WebsocketConfig::default());
-        let (tx, _rx) = again::retry(|| {
-            let websocket_config = Arc::clone(&websocket_config);
-            connect(url.clone().into(), websocket_config)
-        })
-        .await?;
-        Ok(Self { tx })
-    }
-
-    pub async fn attach_app_interface(&mut self, happ_port: u16) -> Result<AdminResponse> {
-        info!(port = ?happ_port, "starting app interface");
-        let msg = AdminRequest::AttachAppInterface {
-            port: Some(happ_port),
-        };
-        self.send(msg).await
-    }
-
-    pub async fn list_app(
-        &mut self,
-        status_filter: Option<AppStatusFilter>,
-    ) -> Result<Vec<InstalledAppId>> {
-        let response = self.send(AdminRequest::ListApps { status_filter }).await?;
-        match response {
-            AdminResponse::AppsListed(info) => {
-                Ok(info.iter().map(|i| i.installed_app_id.to_owned()).collect())
-            }
-            _ => Err(anyhow!("unexpected response: {:?}", response)),
-        }
-    }
-
-    pub async fn list_running_app(&mut self) -> Result<Vec<InstalledAppId>> {
-        let mut running = self.list_app(Some(AppStatusFilter::Running)).await?;
-        let mut enabled = self.list_app(Some(AppStatusFilter::Enabled)).await?;
-        running.append(&mut enabled);
-        Ok(running)
-    }
-
-    pub async fn uninstall_app(&mut self, installed_app_id: &str) -> Result<AdminResponse> {
-        let msg = AdminRequest::UninstallApp {
-            installed_app_id: installed_app_id.to_string(),
-        };
-        self.send(msg).await
-    }
-
-    pub async fn enable_app(&mut self, installed_app_id: &str) -> Result<AdminResponse> {
-        let msg = AdminRequest::EnableApp {
-            installed_app_id: installed_app_id.to_string(),
-        };
-        self.send(msg).await
-    }
-
-    pub async fn disable_app(&mut self, installed_app_id: &str) -> Result<AdminResponse> {
-        let msg = AdminRequest::DisableApp {
-            installed_app_id: installed_app_id.to_string(),
-        };
-        self.send(msg).await
-    }
-
-    async fn send(&mut self, msg: AdminRequest) -> Result<AdminResponse> {
-        let response = self
-            .tx
-            .request(msg)
-            .await
-            .context("failed to send message")?;
-        match response {
-            AdminResponse::Error(error) => Err(anyhow!("error: {:?}", error)),
-            _ => {
-                trace!("send successful");
-                Ok(response)
-            }
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct AppWebsocket {
-    tx: WebsocketSender,
-}
-
-impl AppWebsocket {
-    pub async fn connect(app_port: u16) -> Result<Self> {
-        let url = format!("ws://localhost:{}/", app_port);
-        let url = Url::parse(&url).context("invalid ws:// URL")?;
-        let websocket_config = Arc::new(WebsocketConfig::default());
-        let (tx, _rx) = again::retry(|| {
-            let websocket_config = Arc::clone(&websocket_config);
-            connect(url.clone().into(), websocket_config)
-        })
-        .await?;
-        Ok(Self { tx })
-    }
-
-    pub async fn get_app_info(&mut self, app_id: InstalledAppId) -> Option<AppInfo> {
-        let msg = AppRequest::AppInfo {
-            installed_app_id: app_id,
-        };
-        let response = self.send(msg).await.ok()?;
-        match response {
-            AppResponse::AppInfo(app_info) => app_info,
-            _ => None,
-        }
-    }
-
-    pub async fn zome_call(&mut self, msg: ZomeCall) -> Result<AppResponse> {
-        let app_request = AppRequest::CallZome(Box::new(msg));
-        self.send(app_request).await
-    }
-
-    async fn send(&mut self, msg: AppRequest) -> Result<AppResponse> {
-        let response = self
-            .tx
-            .request(msg)
-            .await
-            .context("failed to send message")?;
-        match response {
-            AppResponse::Error(error) => Err(anyhow!("error: {:?}", error)),
-            _ => {
-                trace!("send successful");
-                Ok(response)
-            }
-        }
-    }
-}
